applications/visualizer/backend/requirements-dev.txt
django-stubs
django_extensions
ipdb
pytest
pytest-django
pytest-asyncio
pytest-unordered
model-bakery
black